19.12.5.1

+++ controller/bg/delete.js
@@ -5,6 +5,7 @@
  * @apiGroup    Bg
  *
  * @apiParam {String} bg_id 项目id
+ * @apiParam {String} userid 用户id
  * @apiSuccessExample Response (example):
  * {
     "code": "0",
@@ -20,14 +21,14 @@ module.exports = (req, send) => {
         send.json({ code, msg, data });
         conn.end();
     }
-    if (!body.bg_id) return toSend("0", "缺少参数");
+    if (!body.bg_id || !body.userid) return toSend("0", "缺少参数");
     conn.connect();
-    conn.query(`delete from bg where bg_id='${body.bg_id}'`,(err,res)=>{
+    conn.query(`delete from bg where bg_id='${body.bg_id}' and userid='${body.userid}'`,(err,res)=>{
         if(err) return toSend("0","系统错误！");
         if(res.affectedRows>0){
             toSend("1","删除成功！");
         }else{
-            toSend("0","没有该bug记录!");
+            toSend("0","没有该bug记录或者没有权限!");
         }
     })
 }
